Refactor application authorization logic in state handlers and replace ApliccationPolicy with ApplicationPolicy
- Updated ApplicationDeleteStateHandler, ApplicationListStateHandler, and ApplicationUpdateStateHandler to include user permission checks before performing actions on applications.
- Removed the outdated ApliccationPolicy and introduced a new ApplicationPolicy to manage authorization for viewing, updating, and deleting applications, ensuring consistent access control across the application.

+++ app/Core/States/ApplicationDeleteStateHandler.php
@@ -32,7 +32,7 @@ class ApplicationDeleteStateHandler extends BaseStateHandler implements StateHan
             $applicationId = $context['application_id_to_delete'];
             if (Str::lower($option) === 'sim') {
                 $application = Application::find($applicationId);
-                if ($application) {
+                if ($application && $user->can('delete', $application)) {
                     $application->delete();
                     $this->evolutionApiService->sendTextMessage($user->phone, __('bot_messages.application_deleted_success'));
                 } else {
